Add MQTT codec
