+++ CMakeLists.txt
@@ -27,6 +27,7 @@ target_include_directories(warp
 target_sources(warp
   PRIVATE
     src/warp/warp.cpp
+    src/warp/mqtt/codec.cpp
     src/warp/mqtt/server.cpp
 )
